pyproject.toml
[project]
name = "NRG-seminar"
version = "0.1.0"
description = "Seminarska naloga pri predmetu NRG"
authors = [{ name = "Iztok Bajcar", email = "iztok.bajcar@gmail.com" }]
dependencies = [
    "laspy[lazrs]>=2.5.4",
    "scikit-learn>=1.6.1",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "pyopengl>=3.1.9",
    "glfw>=2.9.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false

[tool.pdm.scripts]
main = "python3 main.py"
